# src/main/java/com/probestack/forgestudio/design/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.PaymentRequest;
import com.probestack.forgestudio.design.model.RefundRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-13T10:12:31.837361043Z[GMT]")@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /payments/v1/authorize : Authorize payment
     *
     * @param paymentRequest  (required)
     * @return Authorized (status code 200)
     *         or Payment declined (status code 402)
     */
    @Operation(
        operationId = "authorizePayment",
        summary = "Authorize payment",
        responses = {
            @ApiResponse(responseCode = "200", description = "Authorized"),
            @ApiResponse(responseCode = "402", description = "Payment declined")
        },
        security = {
            @SecurityRequirement(name = "apiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payments/v1/authorize",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> authorizePayment(
        @Parameter(name = "PaymentRequest", description = "", required = true) @Valid @RequestBody PaymentRequest paymentRequest
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }


    /**
     * GET /payments/v1/payments/{paymentId} : Get payment status
     *
     * @param paymentId  (required)
     * @return Payment status (status code 200)
     */
    @Operation(
        operationId = "getPayment",
        summary = "Get payment status",
        responses = {
            @ApiResponse(responseCode = "200", description = "Payment status")
        },
        security = {
            @SecurityRequirement(name = "apiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/payments/v1/payments/{paymentId}"
    )
    
    default ResponseEntity<Void> getPayment(
        @Parameter(name = "paymentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentId") String paymentId
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }


    /**
     * POST /payments/v1/payments/{paymentId}/refund : Refund payment
     *
     * @param paymentId  (required)
     * @param refundRequest  (optional)
     * @return Refund initiated (status code 200)
     */
    @Operation(
        operationId = "refundPayment",
        summary = "Refund payment",
        responses = {
            @ApiResponse(responseCode = "200", description = "Refund initiated")
        },
        security = {
            @SecurityRequirement(name = "apiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/payments/v1/payments/{paymentId}/refund",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> refundPayment(
        @Parameter(name = "paymentId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("paymentId") String paymentId,
        @Parameter(name = "RefundRequest", description = "") @Valid @RequestBody(required = false) RefundRequest refundRequest
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }

}